Packing storage values ​​into one to save gas (#32)
* save gas

* Fix test

* fix uint cast

---------

+++ contracts/interfaces/IReClammPool.sol
@@ -11,7 +11,7 @@ struct ReClammPoolParams {
     string version;
     uint256 increaseDayRate;
     uint96 sqrtPriceRatio;
-    uint256 centerednessMargin;
+    uint64 centerednessMargin;
 }
 
 interface IReClammPool is IBasePool {
@@ -27,7 +27,12 @@ interface IReClammPool is IBasePool {
     /// @dev The pool centeredness is too low after a swap.
     error PoolCenterednessTooLow();
 
-    event SqrtPriceRatioUpdated(uint96 startSqrtPriceRatio, uint96 endSqrtPriceRatio, uint32 startTime, uint32 endTime);
+    event SqrtPriceRatioUpdated(
+        uint256 startSqrtPriceRatio,
+        uint256 endSqrtPriceRatio,
+        uint256 startTime,
+        uint256 endTime
+    );
 
     event VirtualBalancesUpdated(uint256[] virtualBalances);
 
@@ -41,7 +46,7 @@ interface IReClammPool is IBasePool {
 
     function getCurrentSqrtPriceRatio() external view returns (uint96);
 
-    function setSqrtPriceRatio(uint96 newSqrtPriceRatio, uint32 startTime, uint32 endTime) external;
+    function setSqrtPriceRatio(uint256 newSqrtPriceRatio, uint256 startTime, uint256 endTime) external;
 
     function setIncreaseDayRate(uint256 newIncreaseDayRate) external;
 }
